pyproject.toml
[project]
name = "omh-plugin"
version = "0.1.0"
description = "Oh My Hermes — plugin infrastructure for OMH skills"
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-cov",
]
